code/src/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.9.2
